Create GOANA.R
This function performs Gene Ontology and KEGG enrichment analysis within the limma package. The user supplies a named list (or list-of-lists) of ENTREZ gene IDs for clusters (or gene subsets) and a vector of ENTREZ gene IDs to be used as a background list (expressed genes). Benjamini-Hochberg false discovery rate (FDR) is calculated. It exports a csv file of the complete enrichment analysis.